Details view with Favorite button working.
