+++ package.json
@@ -67,6 +67,6 @@
     "eslint-plugin-react": "^7.22.0",
     "eslint-plugin-react-hooks": "^4.2.0",
     "eslint-plugin-testing-library": "^3.10.1",
-    "jest": "26.6.0"
+    "jest": "^26.6.0"
   }
 }
