+++ src/components/DetailsView/DetailsView.tsx
@@ -1,4 +1,4 @@
-import React, {useState, useEffect, useContext} from 'react';
+import React, {useState, useEffect} from 'react';
 import {useParams} from 'react-router-dom';
 import styled from 'styled-components';
 import useYouTubeAPI from '../../hooks/useYouTubeAPI';
@@ -7,9 +7,8 @@ import {YouTubeResponse, YouTubeResponseItems} from '../../data-types/YoutubeAPI
 //import bug from './../../data/bug.json';
 //import related_videos from './../../data/related_videos.json';
 import Typography from '@material-ui/core/Typography';
-import Button from '@material-ui/core/Button';
 import RelatedVideo from '../RelatedVideo/RelatedVideo';
-import {GlobalContext} from '../../contexts/GlobalContext';
+import FavoriteButton from '../../components/FavoriteButton/FavoriteButton';
 
 export interface DetailsViewParams {
   videoId: string;
@@ -72,8 +71,6 @@ function DetailsView(props: DetailsViewProps): JSX.Element {
   const [items, setItems] = useState<YouTubeResponseItems[]>([]);
   const [embedHtml, setEmbedHtml] = useState<string>('');
   const [relatedVideos, setRelatedVideos] = useState<JSX.Element[]>([]);
-  const {favoriteVideos, dispatchFav, authenticated} = useContext(GlobalContext);
-  const [buttonLabel, setButtonLabel] = useState<string>('Agregar a favoritos');
 
   useEffect(
     () => {
@@ -130,39 +127,6 @@ function DetailsView(props: DetailsViewProps): JSX.Element {
     };
   }
 
-  function toogleAddFavorite(): void {
-    const id = typeof(items[0].id) === 'string' ? items[0].id : items[0]?.id?.videoId;
-    if (favoriteVideos[`${id}`] && Object.keys(favoriteVideos[`${id}`]).length !== 0) {
-      dispatchFav({
-        type: 'remove',
-        payload: items[0],
-      });
-    }
-    else {
-      dispatchFav({
-        type: 'add',
-        payload: items[0],
-      });
-    }
-  }
-
-  useEffect(
-    () => {
-      const id = items && items[0]
-        ? typeof(items[0].id) === 'string'
-          ? items[0].id
-          : items[0]?.id?.videoId
-        : undefined;
-      if (favoriteVideos[`${id}`] && Object.keys(favoriteVideos[`${id}`]).length !== 0) {
-        setButtonLabel('Remover de favoritos');
-      }
-      else {
-        setButtonLabel('Agregar a favoritos');
-      }
-    },
-    [favoriteVideos, videoId, items]
-  );
-
   return (
     <ParentContainer>
       <RightContainer>
@@ -172,10 +136,7 @@ function DetailsView(props: DetailsViewProps): JSX.Element {
             <Typography gutterBottom variant='h5' component='h2'>
               {items[0]?.snippet?.title}
             </Typography>
-            {
-              authenticated &&
-              <Button onClick={toogleAddFavorite}>{buttonLabel}</Button>
-            }
+            <FavoriteButton item={items[0]} />
           </TitleContainer>
           <JustifyText variant='body2' color='textSecondary'>
             {items[0]?.snippet?.description}
