+++ src/components/Header/components/SearchBox/SearchBox.tsx
@@ -72,6 +72,7 @@ function SearchBox(): JSX.Element {
         value={localSearch}
         onChange={onChange}
         onKeyPress={onKeyPress}
+        id={'search_box'}
       />
     </Search>
   );
